# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Personalizovani Planer API",
        "description": "API dokumentacija za seminar projekat",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Localhost server"
        }
    ],
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Prikaz svih proizvoda",
                "operationId": "f244245293002c3ef578511f93d5b91a",
                "responses": {
                    "200": {
                        "description": "Lista proizvoda"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Dodavanje novog proizvoda",
                "operationId": "df3f418b1f8c0df625ca7b550bb29f4a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "naziv",
                                    "tip",
                                    "opis",
                                    "cena",
                                    "kategorija",
                                    "dostupnaKolicina",
                                    "bojaProizvoda",
                                    "materijalProizvoda"
                                ],
                                "properties": {
                                    "naziv": {
                                        "type": "string"
                                    },
                                    "tip": {
                                        "type": "string"
                                    },
                                    "opis": {
                                        "type": "string"
                                    },
                                    "cena": {
                                        "type": "number"
                                    },
                                    "cenaPopust": {
                                        "type": "number"
                                    },
                                    "kategorija": {
                                        "type": "string"
                                    },
                                    "dostupnaKolicina": {
                                        "type": "integer"
                                    },
                                    "bojaProizvoda": {
                                        "type": "string"
                                    },
                                    "materijalProizvoda": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Proizvod kreiran"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Ažuriranje proizvoda",
                "operationId": "29093e7f8ce57a2bfde2aaaac61f16f1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID proizvoda",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "naziv",
                                    "tip",
                                    "opis",
                                    "cena",
                                    "kategorija",
                                    "dostupnaKolicina",
                                    "bojaProizvoda",
                                    "materijalProizvoda"
                                ],
                                "properties": {
                                    "naziv": {
                                        "type": "string"
                                    },
                                    "tip": {
                                        "type": "string"
                                    },
                                    "opis": {
                                        "type": "string"
                                    },
                                    "cena": {
                                        "type": "number"
                                    },
                                    "cenaPopust": {
                                        "type": "number"
                                    },
                                    "kategorija": {
                                        "type": "string"
                                    },
                                    "dostupnaKolicina": {
                                        "type": "integer"
                                    },
                                    "bojaProizvoda": {
                                        "type": "string"
                                    },
                                    "materijalProizvoda": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Proizvod uspešno ažuriran"
                    },
                    "404": {
                        "description": "Proizvod nije pronađen"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Brisanje proizvoda",
                "operationId": "dba9200f1ffc0d2936f06175893a3989",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID proizvoda",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Proizvod obrisan"
                    },
                    "404": {
                        "description": "Proizvod nije pronađen"
                    }
                }
            }
        }
    },
...
}